add security headers on non-html resources
- Delivers eligible security headers for all resources when possible
- Fix unbuild configuration to allow shared utils between module.ts and runtime
