+++ playground/nuxt.config.ts
@@ -56,7 +56,7 @@ export default defineNuxtConfig({
       interval: 30000,
       headers: true
     },
-    removeLoggers: false
+    removeLoggers: false,
   },
 
   hooks: {
