+++ playground/server/api/test.post.ts
@@ -1,5 +1,5 @@
 export default defineEventHandler(async (event) => {
   console.log('api test', event.path)
   const time = new Date().toISOString()
-  return time
+  return { time }
 })
