+++ src/runtime/composables/nonce.ts
@@ -1,5 +1,5 @@
 import { useNuxtApp } from '#imports'
 
 export function useNonce () {
-  return useNuxtApp().ssrContext?.event?.context.security.nonce
+  return useNuxtApp().ssrContext?.event?.context?.security?.nonce
 }
