+++ src/runtime/nitro/plugins/50-updateCsp.ts
@@ -1,6 +1,6 @@
 import { defineNitroPlugin } from '#imports'
-import { ContentSecurityPolicyValue } from '../../../types/headers'
 import { resolveSecurityRules } from '../context'
+import type { ContentSecurityPolicyValue } from '../../../types/headers'
 
 /**
  * This plugin updates the CSP directives with the nonce and hashes generated by the server.
