+++ src/runtime/nitro/plugins/90-prerenderedHeaders.ts
@@ -1,4 +1,5 @@
 import { defineNitroPlugin, getResponseHeaders, setResponseHeaders, useStorage } from '#imports'
+import { resolveSecurityRules } from '../context'
 
 /**
  * This plugin saves and retrieves prerendered headers in SSG mode.
@@ -8,10 +9,13 @@ export default defineNitroPlugin(async(nitroApp) => {
   if (import.meta.prerender) {
     const prerenderedHeaders: Record<string, Record<string, string>> = {}
     nitroApp.hooks.hook('render:html', (_, { event }) => {
-      // We save the headers for the current path
-      const headers = getResponseHeaders(event) as Record<string, string>
-      const path = event.path.split('?')[0]
-      prerenderedHeaders[path] =  headers
+      const rules = resolveSecurityRules(event)
+      if (rules.enabled && rules.ssg && rules.ssg.nitroHeaders) {
+        // We save the headers for the current path
+        const headers = getResponseHeaders(event) as Record<string, string>
+        const path = event.path.split('?')[0]
+        prerenderedHeaders[path] =  headers
+      }
     })
 
     nitroApp.hooks.hook('close', async() => {
@@ -23,10 +27,13 @@ export default defineNitroPlugin(async(nitroApp) => {
   else {
     const prerenderedHeaders = await useStorage('assets:nuxt-security').getItem<Record<string, Record<string, string>>>('headers.json') || {}
     nitroApp.hooks.hook('beforeResponse', (event) => {
-      const path = event.path.split('?')[0]
-      // We retrieve the headers for the current path
-      if (prerenderedHeaders[path]) {
-        setResponseHeaders(event, prerenderedHeaders[path])
+      const rules = resolveSecurityRules(event)
+      if (rules.enabled && rules.ssg && rules.ssg.nitroHeaders) {
+        const path = event.path.split('?')[0]
+        // We retrieve the headers for the current path
+        if (prerenderedHeaders[path]) {
+          setResponseHeaders(event, prerenderedHeaders[path])
+        }
       }
     })
   }
