+++ src/utils/hash.ts
@@ -1,0 +1,7 @@
+import { createHash } from 'node:crypto';
+
+export function generateHash(content: Buffer | string, hashAlgorithm: string) {
+  const hash = createHash(hashAlgorithm);
+  hash.update(content);
+  return `${hashAlgorithm}-${hash.digest('base64')}`;
+}
