+++ src/utils/merge.ts
@@ -1,0 +1,10 @@
+import { createDefu } from 'defu';
+
+
+export const defuReplaceArray = createDefu((obj, key, value) => {
+  if (Array.isArray(obj[key]) || Array.isArray(value)) {
+    obj[key] = value;
+    return true;
+  }
+});
+
